refactor: created shared components dir and extracted select from ImportPage.jsx into its own component in shared
